refactor: WebUtil 디렉토리 변경 & Component 제거
