+++ src/main/java/com/keeper/homepage/global/util/web/WebUtil.java
@@ -1,16 +1,14 @@
-package com.keeper.homepage.global.util.web.ip;
+package com.keeper.homepage.global.util.web;
 
 
 import jakarta.servlet.http.HttpServletRequest;
 import java.util.Objects;
-import org.springframework.stereotype.Component;
 import org.springframework.web.context.request.RequestContextHolder;
 import org.springframework.web.context.request.ServletRequestAttributes;
 
-@Component
 public class WebUtil {
 
-  public String getUserIP() {
+  public static String getUserIP() {
     HttpServletRequest request = getHttpServletRequestByContext();
 
     String ip = request.getHeader("X-FORWARDED-FOR");
